style: Move paint-order outside of mako
